Move Unbranded to types.ts
